Optimised Blackjack
I created Deck Class and added a string attribute to the Deck Class for differentiation between Player Deck object and Dealer Deck object. For the optimisation, I plan to fully incorporate the Deck Class into the position of the Player Class and remove the Player Class entirely from the Blackjack Project.

I removed all repeated methods between Blackjack Class and Deck Class from the Blackjack Class, and remodififed the Blackjack's attributes, subroutines and/or functions so that it uses the Deck Class instead of its own attributes (that were initially the same as the Deck's which made it inconvenient to program and understand, but now it is simplified).
I have also changed code in the main program (Card Game.cpp) to adjust to the new changes to the Blackjack Class as some subroutine's/function's parameters no longer require arguments and some do.